+++ src/main/java/webserver/header/exception/InvalidPathException.java
@@ -1,4 +1,4 @@
-package webserver.exception;
+package webserver.header.exception;
 
 public class InvalidPathException extends RuntimeException{
     public InvalidPathException(String message) {
