+++ src/main/java/webserver/header/exception/InvalidProtocolException.java
@@ -1,4 +1,4 @@
-package webserver.exception;
+package webserver.header.exception;
 
 public class InvalidProtocolException extends RuntimeException{
     public InvalidProtocolException(String message) {
