+++ src/main/java/webserver/header/exception/InvalidRequestParamsException.java
@@ -1,4 +1,4 @@
-package webserver.exception;
+package webserver.header.exception;
 
 public class InvalidRequestParamsException extends RuntimeException {
     public InvalidRequestParamsException(String message) {
