+++ src/main/java/webserver/header/exception/InvalidUriException.java
@@ -1,4 +1,4 @@
-package webserver.exception;
+package webserver.header.exception;
 
 public class InvalidUriException extends RuntimeException{
     public InvalidUriException(String message) {
