+++ src/main/java/webserver/header/exception/InvalidVersionException.java
@@ -1,4 +1,4 @@
-package webserver.exception;
+package webserver.header.exception;
 
 public class InvalidVersionException extends RuntimeException{
     public InvalidVersionException(String message) {
