+++ src/main/java/webserver/method/exception/InvalidHttpMethodException.java
@@ -1,4 +1,4 @@
-package webserver.exception;
+package webserver.method.exception;
 
 public class InvalidHttpMethodException extends RuntimeException{
     public InvalidHttpMethodException(String message) {
