+++ src/test/java/webserver/header/request/requestline/RequestLineTest.java
@@ -1,4 +1,4 @@
-package webserver.domain;
+package webserver.header.request.requestline;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+
+import webserver.method.HttpMethod;
 
 class RequestLineTest {
