+++ src/test/java/webserver/header/request/requestline/RequestParamsTest.java
@@ -1,4 +1,4 @@
-package webserver.domain;
+package webserver.header.request.requestline;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import webserver.exception.InvalidRequestParamsException;
+import webserver.header.exception.InvalidRequestParamsException;
 
 class RequestParamsTest {
