+++ src/test/java/webserver/header/request/requestline/VersionTest.java
@@ -1,4 +1,4 @@
-package webserver.domain;
+package webserver.header.request.requestline;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import webserver.exception.InvalidVersionException;
+import webserver.header.exception.InvalidVersionException;
 
 class VersionTest {
